lambdifies the traceApertures custom handling to drop out the need for the subclass
